Tidy up. Extend to top slot.
