+++ src/client/bottom-slot.js
@@ -1,6 +1,5 @@
 const oBanner = require('o-banner');
-const { generateMessageEvent, listen } = require('./utils');
-const {messageEventLimitsBreached} = require('./local-tracking');
+const { generateMessageEvent, listen, messageEventLimitsBreached } = require('./utils');
 
 const BOTTOM_SLOT_CONTENT_SELECTOR = '[data-n-messaging-slot="bottom"] [data-n-messaging-component]';
 const BANNER_CLASS = 'n-messaging-banner';
@@ -26,7 +25,6 @@ module.exports = function ({ config={}, guruResult, customSetup }={}) {
 	}
 
 	if (messageEventLimitsBreached(config.name)) {
-		//alert('breach!!!!');
 		trackEventAction('skip');  // todo do we actually need to do this?
 		return;
 	}
