+++ src/client/utils.js
@@ -1,8 +1,19 @@
-const {updateLocalCounter} = require('./local-tracking');
+const cookies = require('js-cookie');
+const _=require('lodash');
+
+const manifest = require('../../manifest');
+const stateCookieName = 'nMessagingEventCounter';
 
 const dispatchEvent = (event) => {
 	document.body.dispatchEvent(event);
 };
+
+const updateLocalCounter =  (messageId, event) => {
+	const currentCounts = cookies.get(stateCookieName) ? JSON.parse(cookies.get(stateCookieName)) : {};
+	const currentCount = _.get(currentCounts, `${messageId}.${event}`) || 0;
+	_.set(currentCounts, `${messageId}.${event}`, currentCount + 1);
+	cookies.set(stateCookieName, currentCounts, { domain: 'ft.com' });
+}
 
 module.exports = {
 	generateMessageEvent: function ({ messageId, position, flag, variant }={}) {
@@ -27,5 +38,17 @@ module.exports = {
 	},
 	listen: function (el, ev, cb) {
 		if (el) el.addEventListener(ev, cb);
+	},
+	messageEventLimitsBreached: function (messageId) {
+		const messageRules = _.get(manifest, `${messageId}.eventRules`);
+		if (!messageRules){
+			return false
+		}
+		const currentCounts = cookies.get(stateCookieName) ? JSON.parse(cookies.get(stateCookieName)) : {};
+
+		return Object.keys(messageRules.maxOccurrences).some( function (eventType) {
+			const eventCount = _.get(currentCounts, `${messageId}.${eventType}`) || 0;
+			return (eventCount >= messageRules.maxOccurrences[eventType]);
+		});
 	}
 };
